+++ imi/src/main/java/org/motechproject/nms/imi/service/impl/IntegrationTestServiceImpl.java
@@ -2,16 +2,33 @@ package org.motechproject.nms.imi.service.impl;
 
 import org.motechproject.nms.imi.repository.FileAuditRecordDataService;
 import org.motechproject.nms.imi.service.IntegrationTestService;
+import org.motechproject.server.config.SettingsFacade;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
-@Service("itService")
+@Service("imiItService")
 public class IntegrationTestServiceImpl implements IntegrationTestService {
+
+    private static final String TESTING_ENVIRONMENT="testing.environment";
 
     @Autowired
     private FileAuditRecordDataService fileAuditRecordDataService;
 
+    /**
+     * SettingsFacade
+     */
+    @Autowired
+    @Qualifier("imiSettings")
+    private SettingsFacade settingsFacade;
+
+
     public void deleteAll() {
+
+        if (!Boolean.parseBoolean(settingsFacade.getProperty(TESTING_ENVIRONMENT))) {
+            throw new IllegalStateException("calling clearDatabase() in a production environment is forbidden!");
+        }
+
         fileAuditRecordDataService.deleteAll();
     }
 }
