+++ region/src/main/java/org/motechproject/nms/region/service/impl/IntegrationTestServiceImpl.java
@@ -13,11 +13,15 @@ import org.motechproject.nms.region.repository.StateDataService;
 import org.motechproject.nms.region.repository.TalukaDataService;
 import org.motechproject.nms.region.repository.VillageDataService;
 import org.motechproject.nms.region.service.IntegrationTestService;
+import org.motechproject.server.config.SettingsFacade;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
-@Service("itService")
+@Service("regionItService")
 public class IntegrationTestServiceImpl implements IntegrationTestService {
+
+    private static final String TESTING_ENVIRONMENT="testing.environment";
 
     @Autowired
     private CircleDataService circleDataService;
@@ -44,8 +48,20 @@ public class IntegrationTestServiceImpl implements IntegrationTestService {
     @Autowired
     private LanguageDataService languageDataService;
 
+    /**
+     * SettingsFacade
+     */
+    @Autowired
+    @Qualifier("regionSettings")
+    private SettingsFacade settingsFacade;
+
 
     public void deleteAll() {
+
+        if (!Boolean.parseBoolean(settingsFacade.getProperty(TESTING_ENVIRONMENT))) {
+            throw new IllegalStateException("calling clearDatabase() in a production environment is forbidden!");
+        }
+
         circleDataService.deleteAll();
         districtDataService.deleteAll();
         healthBlockDataService.deleteAll();
